Add ready, kicking, and an http admin page.
This brings the dedicated server to feature parity with all changes from merging in the 'HostAbilities-ReadyCheck' branch (ad3b5c9). Since there is no lua-based host, kicking players is managed by an http management page on the dedicated server. This required switching to a listener pattern for the playerlist, which (somewhat) simplified logic in room.go.