Make all luasocket I/O non-blocking
Previously, sync.syncconfig waited for several messages to be sent and received without yielding. As a result, BizHawk can block for a second or two between frames. This experience isn't bad when only connecting once, but it's awful if someone's trying to play while reconnects attempts are being made in the background.

Tested:
Verified creating a room, joining a room, and sending items still works.