Cache sync hashes to avoid extra recomputation
Computing the sync hashes is relatively slow (a few hunder ms) due to the disk I/O. Caching the hash reduces performance issues on the server when clients are connecting and on the client when reconnecting. The sync hashes are cached per ramcontroller since it's simpler than clearing the sync hash whenever the active ramcontroller changes.

require() is especially slow since it requires interpreting the lua file; performing it once on startup is a significant performance improvement.

Also switch to concatenating the lines all at once to reduce the number of memory reallocations.

Tested:
Created a room and had a client connect several times.