Default to the ramcontroller that matches the current ROM
Each ramcontroller may provide a Lua pattern to match against the ROMs name (as provided by gameinfo.getromname()); the ramcontroller with the first matching pattern is used as the default.

The pattern is specified by a comment rather than a function or global variable since (a) it's slow to run all of the ramcontrollers on startup, and (b) the ramcontrollers may modify global state. A comment is also trivially backwards-compatible with any existing third-party ramcontrollers.